Add array-constructor and sizeof support
Introduce AST node ArrayContructor and AnyArray to support `[for N => expr]` array constructors and a `.sizeof` type operator. Parser updates parse array constructors and treat `.sizeof` as a type-only field access; added utility rename current_is_keyword and LAMBDA_ARROW token. AST display, name-resolver, and resolver updated to collect/resolve new nodes. HIR changes: add Sizeof expression, desugar array constructors into literal arrays, and create an internal ___Array struct for array field access. Typed HIR updates add u32/sizeof tracking, improved numeric/binop type checks, and various type helper additions (pointer/generic checks, modifier handling). Misc: TypesMap/ThirTypesMap and related maps updated to include array_struct and default handling, plus small refactors and error messaging improvements.

+++ soul_tester/soul_src/soul_clib/clib.h
@@ -28,7 +28,7 @@ typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
 typedef uint64_t u64;
-typedef unsigned int uint;
+typedef size_t uint;
 
 typedef float f32;
 typedef double f64;
